requirements.txt
python-docx
gradio
chromadb
langchain
openai